@@ -1,6 +1,6 @@
 {
   "name": "react-dfp",
-  "version": "0.3.8",
+  "version": "0.4.2",
   "homepage": "https://github.com/jaanauati/react-dfp/",
   "author": {
     "name": "Jonatan Alexis Anauati",
@@ -28,61 +28,61 @@
   ],
   "main": "lib/index.js",
   "peerDependencies": {
-    "react": ">=0.14.0",
+    "react": ">=0.15.6",
     "react-dom": ">=0.14.0"
   },
   "devDependencies": {
     "array.from": "^1.0.1",
-    "babel": "^6.5.2",
+    "babel-cli": "^6.5.2",
+    "babel-eslint": "^7.2.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.8.0",
     "babel-plugin-transform-react-inline-elements": "^6.8.0",
     "babel-polyfill": "^6.7.4",
-    "babel-preset-airbnb": "^1.1.1",
+    "babel-preset-airbnb": "^2.2.3",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
     "babelify": "^7.3.0",
     "brfs": "^1.4.2",
-    "browser-sync": "^1.8.3",
-    "browserify": "^9.0.3",
-    "browserify-istanbul": "^0.2.1",
+    "browser-sync": "^2.18.8",
+    "browserify": "^14.3.0",
+    "browserify-istanbul": "^2.0.0",
     "browserify-shim": "^3.8.0",
-    "chai": "^2.2.0",
+    "chai": "^3.5.0",
     "escope": "^3.3.0",
-    "eslint": "^2.10.0",
-    "eslint-config-airbnb": "^9.0.1",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^5.1.1",
+    "eslint-plugin-react": "^6.10.3",
     "estraverse": "^4.2.0",
-    "grunt": "0.4.x",
+    "grunt": "1.0.1",
     "grunt-babel": "^6.0.0",
     "grunt-cli": "^1.2.0",
-    "grunt-contrib-clean": "^0.4.1",
-    "grunt-eslint": "^18.1.0",
-    "grunt-karma": "^1.0.0",
+    "grunt-contrib-clean": "^1.0.0",
+    "grunt-eslint": "^19.0.0",
+    "grunt-karma": "^2.0.0",
     "jasmine-core": "^2.4.1",
-    "karma": "^0.13.22",
+    "jsdom": "^9.12.0",
+    "karma": "^1.6.0",
     "karma-babel-preprocessor": "^6.0.1",
-    "karma-browserify": "^2.0.0",
+    "karma-browserify": "^5.1.1",
     "karma-chai": "^0.1.0",
     "karma-jasmine": "^1.0.2",
-    "karma-jsdom-launcher": "^3.0.0",
-    "karma-mocha": "^0.1.10",
-    "karma-mocha-reporter": "^0.3.1",
-    "karma-script-launcher": "~0.1.0",
+    "karma-jsdom-launcher": "^6.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.3",
+    "karma-script-launcher": "^1.0.0",
     "karma-sinon": "^1.0.4",
-    "load-grunt-config": "^0.7.2",
-    "mocha": "^1.18.2",
-    "react": "^15.0.2",
+    "load-grunt-config": "^0.19.2",
+    "mocha": "^3.2.0",
+    "react": "^15.6.0",
     "react-addons-test-utils": "^15.0.2",
     "react-dom": "^15.0.2",
-    "sinon": "^1.17.4",
+    "sinon": "^2.1.0",
     "sinon-chai": "^2.6.0",
-    "time-grunt": "^0.2.10",
-    "watchify": "^2.4.0"
-  },
-  "engines": {
-    "node": "<=6.0.0"
+    "time-grunt": "^1.4.0",
+    "watchify": "^3.9.0"
   },
   "licenses": [
     {
@@ -94,5 +94,9 @@
     "plugins": [
       "transform-object-rest-spread"
     ]
+  },
+  "dependencies": {
+    "babel-plugin-transform-object-assign": "^6.22.0",
+    "prop-types": "^15.5.6"
   }
 }
