@@ -2,5 +2,5 @@ module.exports = {
     options: {
         config: '.eslintrc.json'
     },
-    all: [ './js/*.js', './spec/*.js' ]
+    all: [ './js/*.js' ]
 };
