@@ -1,3 +1,7 @@
-export { DFPManager } from './manager';
-export { AdSlot } from './adslot';
-export { DFPSlotsProvider } from './dfpslotsprovider';
+import manager from './manager';
+import slot from './adslot';
+import provider from './dfpslotsprovider';
+
+export const DFPManager = manager;
+export const AdSlot = slot;
+export const DFPSlotsProvider = provider;
