@@ -3,30 +3,22 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+exports.DFPSlotsProvider = exports.AdSlot = exports.DFPManager = undefined;
 
 var _manager = require('./manager');
 
-Object.defineProperty(exports, 'DFPManager', {
-  enumerable: true,
-  get: function get() {
-    return _manager.DFPManager;
-  }
-});
+var _manager2 = _interopRequireDefault(_manager);
 
 var _adslot = require('./adslot');
 
-Object.defineProperty(exports, 'AdSlot', {
-  enumerable: true,
-  get: function get() {
-    return _adslot.AdSlot;
-  }
-});
+var _adslot2 = _interopRequireDefault(_adslot);
 
 var _dfpslotsprovider = require('./dfpslotsprovider');
 
-Object.defineProperty(exports, 'DFPSlotsProvider', {
-  enumerable: true,
-  get: function get() {
-    return _dfpslotsprovider.DFPSlotsProvider;
-  }
-});
+var _dfpslotsprovider2 = _interopRequireDefault(_dfpslotsprovider);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var DFPManager = exports.DFPManager = _manager2.default;
+var AdSlot = exports.AdSlot = _adslot2.default;
+var DFPSlotsProvider = exports.DFPSlotsProvider = _dfpslotsprovider2.default;
