@@ -3,15 +3,14 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.DFPManager = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
+var _events = require('events');
 
 var _utils = require('./utils');
 
 var Utils = _interopRequireWildcard(_utils);
-
-var _events = require('events');
 
 function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
@@ -20,14 +19,22 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 var loadAlreadyCalled = false;
 var googleGPTScriptLoadPromise = null;
 var registeredSlots = {};
-var loadedGptSlots = {};
-var pubadsService = null;
 var managerAlreadyInitialized = false;
 var globalTargetingArguments = {};
 
-var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), {
+var DFPManager = _extends(new _events.EventEmitter().setMaxListeners(0), {
   setTargetingArguments: function setTargetingArguments(data) {
-    Object.assign(globalTargetingArguments, data);
+    _extends(globalTargetingArguments, data);
+    if (managerAlreadyInitialized === true) {
+      this.getGoogletag().then(function (googletag) {
+        googletag.cmd.push(function () {
+          var pubadsService = googletag.pubads();
+          Object.keys(globalTargetingArguments).forEach(function (varName) {
+            pubadsService.setTargeting(varName, globalTargetingArguments[varName]);
+          });
+        });
+      });
+    }
   },
   getTargetingArguments: function getTargetingArguments() {
     return _extends({}, globalTargetingArguments);
@@ -47,16 +54,14 @@ var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), 
       managerAlreadyInitialized = true;
       this.getGoogletag().then(function (googletag) {
         googletag.cmd.push(function () {
-          pubadsService = googletag.pubads();
+          var pubadsService = googletag.pubads();
           pubadsService.addEventListener('slotRenderEnded', function (event) {
             var slotId = event.slot.getSlotElementId();
             _this.emit('slotRenderEnded', { slotId: slotId, event: event });
           });
           var targetingArguments = _this.getTargetingArguments();
           Object.keys(targetingArguments).forEach(function (varName) {
-            if (targetingArguments.hasOwnProperty(varName)) {
-              pubadsService.setTargeting(varName, targetingArguments[varName]);
-            }
+            pubadsService.setTargeting(varName, targetingArguments[varName]);
           });
         });
       });
@@ -68,11 +73,15 @@ var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), 
     }
     return googleGPTScriptLoadPromise;
   },
+  setCollapseEmptyDivs: function setCollapseEmptyDivs(collapse) {
+    this.collapseEmptyDivs = collapse;
+  },
   load: function load(slotId) {
     var _this2 = this;
 
     this.init();
     var availableSlots = {};
+
     if (loadAlreadyCalled === true) {
       var slot = registeredSlots[slotId];
       if (slot !== undefined) {
@@ -85,51 +94,49 @@ var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), 
     availableSlots = Object.keys(availableSlots).filter(function (id) {
       return !availableSlots[id].loading;
     }).reduce(function (result, id) {
-      return Object.assign(result, _defineProperty({}, id, Object.assign(availableSlots[id], { loading: true })));
+      return _extends(result, _defineProperty({}, id, _extends(availableSlots[id], { loading: true })));
     }, {});
     this.getGoogletag().then(function (googletag) {
       Object.keys(availableSlots).forEach(function (currentSlotId) {
         availableSlots[currentSlotId].loading = false;
 
-        if (!loadedGptSlots[currentSlotId]) {
-          loadedGptSlots[currentSlotId] = true;
-          googletag.cmd.push(function () {
-            var slot = availableSlots[currentSlotId];
-            var gptSlot = void 0;
-            var adUnit = slot.dfpNetworkId + '/' + slot.adUnit;
-            if (slot.renderOutOfThePage === true) {
-              gptSlot = googletag.defineOutOfPageSlot(adUnit, currentSlotId);
-            } else {
-              gptSlot = googletag.defineSlot(adUnit, slot.sizes, currentSlotId);
-            }
-            slot.gptSlot = gptSlot;
-            var slotTargetingArguments = _this2.getSlotTargetingArguments(currentSlotId);
-            if (slotTargetingArguments !== null) {
-              Object.keys(slotTargetingArguments).forEach(function (varName) {
-                if (slotTargetingArguments.hasOwnProperty(varName)) {
-                  slot.gptSlot.setTargeting(varName, slotTargetingArguments[varName]);
-                }
-              });
-            }
-            slot.gptSlot.addService(googletag.pubads());
-            if (slot.sizeMapping) {
-              var smbuilder = googletag.sizeMapping();
-              slot.sizeMapping.forEach(function (mapping) {
-                smbuilder = smbuilder.addSize(mapping.viewport, mapping.sizes);
-              });
-              slot.gptSlot.defineSizeMapping(smbuilder.build());
-            }
-          });
-        }
+        googletag.cmd.push(function () {
+          var slot = availableSlots[currentSlotId];
+          var gptSlot = void 0;
+          var adUnit = slot.dfpNetworkId + '/' + slot.adUnit;
+          if (slot.renderOutOfThePage === true) {
+            gptSlot = googletag.defineOutOfPageSlot(adUnit, currentSlotId);
+          } else {
+            gptSlot = googletag.defineSlot(adUnit, slot.sizes, currentSlotId);
+          }
+          slot.gptSlot = gptSlot;
+          var slotTargetingArguments = _this2.getSlotTargetingArguments(currentSlotId);
+          if (slotTargetingArguments !== null) {
+            Object.keys(slotTargetingArguments).forEach(function (varName) {
+              slot.gptSlot.setTargeting(varName, slotTargetingArguments[varName]);
+            });
+          }
+          slot.gptSlot.addService(googletag.pubads());
+          if (slot.sizeMapping) {
+            var smbuilder = googletag.sizeMapping();
+            slot.sizeMapping.forEach(function (mapping) {
+              smbuilder = smbuilder.addSize(mapping.viewport, mapping.sizes);
+            });
+            slot.gptSlot.defineSizeMapping(smbuilder.build());
+          }
+        });
       });
 
       googletag.cmd.push(function () {
         googletag.pubads().enableSingleRequest();
+
+        if (_this2.collapseEmptyDivs === true || _this2.collapseEmptyDivs === false) {
+          googletag.pubads().collapseEmptyDivs(_this2.collapseEmptyDivs);
+        }
+
         googletag.enableServices();
-        Object.keys(availableSlots).forEach(function (_slotId) {
-          if (availableSlots.hasOwnProperty(_slotId)) {
-            googletag.display(_slotId);
-          }
+        Object.keys(availableSlots).forEach(function (theSlotId) {
+          googletag.display(theSlotId);
         });
       });
     });
@@ -173,10 +180,17 @@ var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), 
         slotId = _ref.slotId,
         slotShouldRefresh = _ref.slotShouldRefresh;
 
-    if (!registeredSlots.hasOwnProperty(slotId)) {
-      registeredSlots[slotId] = { slotId: slotId, sizes: sizes, renderOutOfThePage: renderOutOfThePage,
-        dfpNetworkId: dfpNetworkId, adUnit: adUnit, targetingArguments: targetingArguments,
-        sizeMapping: sizeMapping, slotShouldRefresh: slotShouldRefresh, loading: false
+    if (!Object.prototype.hasOwnProperty.call(registeredSlots, slotId)) {
+      registeredSlots[slotId] = {
+        slotId: slotId,
+        sizes: sizes,
+        renderOutOfThePage: renderOutOfThePage,
+        dfpNetworkId: dfpNetworkId,
+        adUnit: adUnit,
+        targetingArguments: targetingArguments,
+        sizeMapping: sizeMapping,
+        slotShouldRefresh: slotShouldRefresh,
+        loading: false
       };
     }
     if (loadAlreadyCalled === true) {
@@ -198,3 +212,5 @@ var DFPManager = exports.DFPManager = Object.assign(new _events.EventEmitter(), 
     this.removeListener('slotRenderEnded', cb);
   }
 });
+
+exports.default = DFPManager;
