@@ -2,7 +2,7 @@ import 'babel-polyfill';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import TestUtils from 'react-addons-test-utils';
+import ReactTestUtils from 'react-dom/test-utils';
 import { expect } from 'chai';
 import sinon from 'sinon';
 
@@ -13,19 +13,25 @@ describe('AdSlot', () => {
     const compProps = {
       dfpNetworkId: '1000',
       adUnit: 'foo/bar/baz',
-      slotId: 'testElement',
       sizes: [[728, 90], 'fluid'],
     };
+    const compTwoProps = {
+      ...compProps,
+      slotId: 'testElement',
+    };
 
-    let component;
-    beforeEach(() => {
-      component = TestUtils.renderIntoDocument(<AdSlot {...compProps} />);
+    it('renders an AdSlot with the given elementId', () => {
+      const component = ReactTestUtils.renderIntoDocument(<AdSlot {...compTwoProps} />);
+      const box = ReactTestUtils.findRenderedDOMComponentWithClass(component, 'adBox');
+      expect(box.id).to.equal('testElement');
     });
 
-    it('renders an adBox with the given elementId', () => {
-      const box = TestUtils.findRenderedDOMComponentWithClass(component,
-                                                              'adBox');
-      expect(box.id).to.equal('testElement');
+    it('renders two AdSlots and verify that those get different ids', () => {
+      let componentOne = ReactTestUtils.renderIntoDocument(<AdSlot {...compProps} />);
+      let componentTwo = ReactTestUtils.renderIntoDocument(<AdSlot {...compProps} />);
+      componentOne = ReactTestUtils.findRenderedDOMComponentWithClass(componentOne, 'adBox');
+      componentTwo = ReactTestUtils.findRenderedDOMComponentWithClass(componentTwo, 'adBox');
+      expect(componentOne.id).to.not.equal(componentTwo.id);
     });
   });
 
@@ -43,9 +49,7 @@ describe('AdSlot', () => {
         sizes: [[728, 90]],
       };
 
-      TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
-      );
+      ReactTestUtils.renderIntoDocument(<AdSlot {...compProps} />);
 
       sinon.assert.calledOnce(DFPManager.registerSlot);
       sinon.assert.calledWithMatch(DFPManager.registerSlot, compProps);
@@ -59,9 +63,7 @@ describe('AdSlot', () => {
         sizes: [[728, 90]],
       };
 
-      TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
-      );
+      ReactTestUtils.renderIntoDocument(<AdSlot {...compProps} />);
 
       expect(DFPManager.getRefreshableSlots()).to.contain.all.keys([compProps.slotId]);
       expect(DFPManager.getRefreshableSlots()[compProps.slotId]).to.contain.all.keys(compProps);
@@ -76,8 +78,8 @@ describe('AdSlot', () => {
         shouldRefresh: () => false,
       };
 
-      TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
+      ReactTestUtils.renderIntoDocument(
+        <AdSlot {...compProps} />,
       );
       expect(Object.keys(DFPManager.getRefreshableSlots()).length).to.equal(0);
     });
@@ -91,8 +93,8 @@ describe('AdSlot', () => {
         targetingArguments: { team: 'river plate', player: 'pisculichi' },
       };
 
-      TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
+      ReactTestUtils.renderIntoDocument(
+        <AdSlot {...compProps} />,
       );
       expect(DFPManager.getSlotTargetingArguments(compProps.slotId))
         .to.contain.all.keys(compProps.targetingArguments);
@@ -106,8 +108,8 @@ describe('AdSlot', () => {
         sizes: [[728, 90]],
       };
 
-      TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
+      ReactTestUtils.renderIntoDocument(
+        <AdSlot {...compProps} />,
       );
       expect(DFPManager.getSlotTargetingArguments(compProps.slotId)).to.equal(null);
     });
@@ -121,8 +123,8 @@ describe('AdSlot', () => {
         sizes: [[728, 90]],
       };
 
-      const component = TestUtils.renderIntoDocument(
-        <AdSlot {...compProps} />
+      const component = ReactTestUtils.renderIntoDocument(
+        <AdSlot {...compProps} />,
       );
 
       ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(component).parentNode);
